upper_bound_benchmark/requirements.txt
# Runtime
gymnasium==0.29.1
ale-py==0.8.1
AutoROM==0.4.2
AutoROM.accept-rom-license==0.6.1
numpy>=1.24,<3.0
torch>=2.0.0,<3.0.0
Pillow>=9.5.0

# Optional for rendering (comment out if headless)
pygame>=2.5.0
